yt_data_v1_v2/requirements.txt
google-api-python-client==2.104.0
youtube-transcript-api==0.6.1
pandas==2.0.3
pyarrow==13.0.0
isodate==0.6.1
python-dotenv==1.0.0